requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sentence-transformers==2.2.2
chromadb==0.4.18
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-generativeai==0.3.2
ollama==0.1.7
pydantic==2.5.0
